face-id20.01.24new - Copy/requirements.txt
Flask==2.3.2
numpy==1.26.4
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
Pillow==10.0.1
Pillow==10.3.0
